tests/wpt/tests/css/css-text/parsing/text-wrap-style-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-wrap-style with valid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap-style">
<meta name="assert" content="text-wrap-style supports the full grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-wrap-style", "auto");
test_valid_value("text-wrap-style", "balance");
test_valid_value("text-wrap-style", "pretty");
test_valid_value("text-wrap-style", "stable");

test_valid_value("text-wrap-style", "initial");
test_valid_value("text-wrap-style", "inherit");
test_valid_value("text-wrap-style", "unset");
test_valid_value("text-wrap-style", "revert");
test_valid_value("text-wrap-style", "revert-layer");
</script>
</body>
</html>
