third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-text/parsing/text-wrap-style-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['text-wrap-style'] = "auto" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "balance" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "pretty" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "stable" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "initial" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "inherit" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "unset" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "revert" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "revert-layer" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
